+++ lib/src/num_extensions.dart
@@ -1,6 +1,3 @@
-DateTime toUnix(int timestamp) =>
-    DateTime.fromMicrosecondsSinceEpoch(timestamp);
-
 extension NumExtension on num {
   bool isBetween(num min, num max, {bool orEqualSomeEdge = false}) {
     if (orEqualSomeEdge) {